+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/VersusBuyAndHoldCriterionTest.java
@@ -37,7 +37,7 @@ import static org.ta4j.core.num.NaN.NaN;
 public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest{
 
     public VersusBuyAndHoldCriterionTest(Function<Number, Num> numFunction) {
-        super(numFunction);
+        super((params) -> new VersusBuyAndHoldCriterion((AnalysisCriterion) params[0]), numFunction);
     }
 
     @Test
@@ -47,7 +47,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest{
                 Order.buyAt(0,series), Order.sellAt(2,series),
                 Order.buyAt(3,series), Order.sellAt(5,series));
 
-        AnalysisCriterion buyAndHold = new VersusBuyAndHoldCriterion(new TotalProfitCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new TotalProfitCriterion());
         assertNumEquals(1.10 * 1.05 / 1.05, buyAndHold.calculate(series, tradingRecord));
     }
 
@@ -58,7 +58,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest{
                 Order.buyAt(0,series), Order.sellAt(1,series),
                 Order.buyAt(2,series), Order.sellAt(5,series));
 
-        AnalysisCriterion buyAndHold = new VersusBuyAndHoldCriterion(new TotalProfitCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new TotalProfitCriterion());
         assertNumEquals(0.95 * 0.7 / 0.7, buyAndHold.calculate(series, tradingRecord));
     }
 
@@ -67,7 +67,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest{
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 95, 100, 80, 85, 70);
         Trade trade = new Trade(Order.buyAt(0,series), Order.sellAt(1,series));
 
-        AnalysisCriterion buyAndHold = new VersusBuyAndHoldCriterion(new TotalProfitCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new TotalProfitCriterion());
         assertNumEquals((100d / 70) / (100d / 95), buyAndHold.calculate(series, trade));
     }
 
@@ -75,7 +75,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest{
     public void calculateWithNoTrades() {
         MockTimeSeries series = new MockTimeSeries(numFunction, 100, 95, 100, 80, 85, 70);
 
-        AnalysisCriterion buyAndHold = new VersusBuyAndHoldCriterion(new TotalProfitCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new TotalProfitCriterion());
         assertNumEquals(1 / 0.7, buyAndHold.calculate(series, new BaseTradingRecord()));
     }
 
@@ -86,7 +86,7 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest{
                 Order.buyAt(0, NaN, NaN), Order.sellAt(1, NaN, NaN),
                 Order.buyAt(2, NaN, NaN), Order.sellAt(5, NaN, NaN));
 
-        AnalysisCriterion buyAndHold = new VersusBuyAndHoldCriterion(new AverageProfitCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new AverageProfitCriterion());
 
         assertNumEquals(Math.pow(95d/100 * 130d/100, 1d/6) / Math.pow(130d / 100, 1d/6), buyAndHold.calculate(series, tradingRecord));
     }
@@ -98,14 +98,14 @@ public class VersusBuyAndHoldCriterionTest extends AbstractCriterionTest{
                 Order.buyAt(0,series), Order.sellAt(1,series),
                 Order.buyAt(2,series), Order.sellAt(5,series));
 
-        AnalysisCriterion buyAndHold = new VersusBuyAndHoldCriterion(new NumberOfBarsCriterion());
+        AnalysisCriterion buyAndHold = getCriterion(new NumberOfBarsCriterion());
 
         assertNumEquals(6d/6d, buyAndHold.calculate(series, tradingRecord));
     }
 
     @Test
     public void betterThan() {
-        AnalysisCriterion criterion = new VersusBuyAndHoldCriterion(new TotalProfitCriterion());
+        AnalysisCriterion criterion = getCriterion(new TotalProfitCriterion());
         assertTrue(criterion.betterThan(numOf(2.0), numOf(1.5)));
         assertFalse(criterion.betterThan(numOf(1.5), numOf(2.0)));
     }